requirements.txt
zmq
numpy
pandas
scikit-learn
python-xlib
Pillow
whereami
